Reorganize code by @pzmarzly's suggestion
- Event aliases moved to events.py
- Networking and webserver-related functions moved to network.py
- A couple of extraneous lines deleted
