+++ bin/yoke
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 import yoke
-from yoke import EVENTS
 import argparse
 import sys
 import errno
