+++ yoke/events.py
@@ -593,3 +593,63 @@ ABS_MT_DISTANCE = (0x03, 0x3b)
 ABS_MT_TOOL_X = (0x03, 0x3c)
 ABS_MT_TOOL_Y = (0x03, 0x3d)
 ABS_MAX = (0x03, 0x3f)
+
+ALIAS = {
+'j1': 'ABS_X,ABS_Y',
+'j2': 'ABS_RX,ABS_RY',
+'j3': 'ABS_TOOL_WIDTH,ABS_MAX',
+'s1': 'ABS_X,ABS_Y',
+'s2': 'ABS_RX,ABS_RY',
+'s3': 'ABS_TOOL_WIDTH,ABS_MAX',
+'mx': 'ABS_MISC',
+'my': 'ABS_RZ',
+'mz': 'ABS_Z',
+'ma': 'ABS_TILT_X',
+'mb': 'ABS_WHEEL',
+'mg': 'ABS_TILT_Y',
+'pa': 'ABS_GAS',
+'pb': 'ABS_BRAKE',
+'pt': 'ABS_THROTTLE',
+'k1': 'ABS_VOLUME',
+'k2': 'ABS_RUDDER',
+'k3': 'ABS_PRESSURE',
+'k4': 'ABS_DISTANCE',
+'a1': 'ABS_HAT0X',
+'a2': 'ABS_HAT0Y',
+'a3': 'ABS_HAT1X',
+'a4': 'ABS_HAT1Y',
+'a5': 'ABS_HAT2X',
+'a6': 'ABS_HAT2Y',
+'a7': 'ABS_HAT3X',
+'a8': 'ABS_HAT3Y',
+'bs': 'BTN_START',
+'bg': 'BTN_SELECT',
+'bm': 'BTN_MODE',
+'b1': 'BTN_GAMEPAD',
+'b2': 'BTN_EAST',
+'b3': 'BTN_WEST',
+'b4': 'BTN_NORTH',
+'b5': 'BTN_TL',
+'b6': 'BTN_TR',
+'b7': 'BTN_TL2',
+'b8': 'BTN_TR2',
+'b9': 'BTN_A',
+'b10': 'BTN_B',
+'b11': 'BTN_C',
+'b12': 'BTN_X',
+'b13': 'BTN_Y',
+'b14': 'BTN_Z',
+'b15': 'BTN_TOP',
+'b16': 'BTN_TOP2',
+'b17': 'BTN_PINKIE',
+'b18': 'BTN_BASE',
+'b19': 'BTN_BASE2',
+'b20': 'BTN_BASE3',
+'b21': 'BTN_BASE4',
+'b22': 'BTN_BASE5',
+'b23': 'BTN_BASE6',
+'b24': 'BTN_THUMB',
+'b25': 'BTN_THUMB2',
+'b26': 'BTN_TRIGGER',
+'dp': 'BTN_DPAD_UP,BTN_DPAD_LEFT,BTN_DPAD_DOWN,BTN_DPAD_RIGHT',
+}
